Linux compat fixed 1 - Fix a case-sensitivity issue fir files/directories on Linux - Fix a float-to-int conversion issue - Fix an issue where the settings folder might not exist

+++ SubLink.Discord/Services/DiscordService+Actions.cs
@@ -70,8 +70,8 @@ internal sealed partial class DiscordService {
     public void SelectTextChannel(string channelId) =>
         _discord?.SendCommand(1, DiscordIpcMessage.SelectTextChannel(channelId));
 
-    public void SetUserVolume(string userId, float vol) =>
-        _discord?.SendCommand(1, DiscordIpcMessage.SetUserVoiceSettings(userId, volume: (int)vol));
+    public void SetUserVolume(string userId, int vol) =>
+        _discord?.SendCommand(1, DiscordIpcMessage.SetUserVoiceSettings(userId, volume: vol));
 
     public void SetUserMute(string userId, bool mute) =>
         _discord?.SendCommand(1, DiscordIpcMessage.SetUserVoiceSettings(userId, mute: mute));
